requirements.txt
alpaca-py==0.37.0
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
msgpack==1.1.0
numpy==2.2.2
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
sseclient-py==1.8.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
alpaca-trade-api
tabulate
python-dotenv
